scenarios/events/app/requirements.txt
fastapi
uvicorn
confluent_kafka
python-dotenv==0.10.1
